Full support of aws-secret-manager
