+++ docs/modules/ROOT/examples/components/aws-secrets-manager.yml
@@ -2,11 +2,11 @@
 # This file was generated by camel-quarkus-maven-plugin:update-extension-doc-page
 cqArtifactId: camel-quarkus-aws-secrets-manager
 cqArtifactIdBase: aws-secrets-manager
-cqNativeSupported: false
-cqStatus: Preview
+cqNativeSupported: true
+cqStatus: Stable
 cqDeprecated: false
 cqJvmSince: 2.0.0
-cqNativeSince: n/a
+cqNativeSince: 3.19.0
 cqCamelPartName: aws-secrets-manager
 cqCamelPartTitle: AWS Secrets Manager
 cqCamelPartDescription: Manage secrets using AWS Secrets Manager.
