+++ extensions/aws-secrets-manager/pom.xml
@@ -18,34 +18,20 @@
 
 -->
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>org.apache.camel.quarkus</groupId>
-        <artifactId>camel-quarkus-integration-test-groups</artifactId>
+        <artifactId>camel-quarkus-extensions</artifactId>
         <version>3.19.0-SNAPSHOT</version>
+        <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <artifactId>camel-quarkus-integration-tests-aws2</artifactId>
+    <artifactId>camel-quarkus-aws-secrets-manager-parent</artifactId>
+    <name>Camel Quarkus :: AWS Secrets Manager</name>
     <packaging>pom</packaging>
 
-    <name>Camel Quarkus :: AWS 2 Integration Tests</name>
-
-    <properties>
-        <quarkus.banner.enabled>false</quarkus.banner.enabled>
-    </properties>
-
     <modules>
-        <!-- extensions a..z; do not remove this comment, it is important when sorting via  mvn process-resources -Pformat -->
-        <module>aws-secrets-manager</module>
-        <module>aws2-cw</module>
-        <module>aws2-ddb</module>
-        <module>aws2-kinesis</module>
-        <module>aws2-lambda</module>
-        <module>aws2-s3</module>
-        <module>aws2-ses</module>
-        <module>aws2-sqs</module>
-        <module>aws2-sqs-sns</module>
+        <module>deployment</module>
+        <module>runtime</module>
     </modules>
-
 </project>
