+++ integration-test-groups/aws2/aws-secrets-manager/src/main/resources/application.properties
