+++ integration-test-groups/aws2/aws-secrets-manager/src/test/resources/META-INF/services/org.apache.camel.quarkus.test.support.aws2.Aws2TestEnvCustomizer
